requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
orjson
motor
beanie
chem_templates